refactor: 중복 코드 Callback, Executor로 분리
