+++ jdbc/src/main/java/org/springframework/jdbc/core/JdbcTemplate.java
@@ -1,66 +1,38 @@
 package org.springframework.jdbc.core;
 
-import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
-import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.List;
 import javax.sql.DataSource;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.dao.DataAccessException;
 import org.springframework.dao.IncorrectResultSizeException;
 import org.springframework.dao.ResultEmptyException;
-import org.springframework.jdbc.datasource.DataSourceUtils;
 
 public class JdbcTemplate {
 
-    private static final Logger log = LoggerFactory.getLogger(JdbcTemplate.class);
-
-    private final DataSource dataSource;
+    private final PreparedStatementExecutor preparedStatementExecutor;
 
     public JdbcTemplate(final DataSource dataSource) {
-        this.dataSource = dataSource;
+        this.preparedStatementExecutor = new PreparedStatementExecutor(dataSource);
     }
 
     public void update(final String sql, final Object... args) {
-        try (final Connection connection = DataSourceUtils.getConnection(dataSource);
-             final PreparedStatement pstmt = connection.prepareStatement(sql)) {
-            log.debug("query : {}", sql);
-
-            setPreparedStatementArguments(pstmt, args);
-            pstmt.executeUpdate();
-        } catch (SQLException e) {
-            log.error(e.getMessage(), e);
-            throw new DataAccessException(e);
-        }
+        preparedStatementExecutor.execute(sql, PreparedStatement::executeUpdate, args);
     }
 
-    public <T> List<T> query(final String sql, final RowMapper<T> rowMapper) {
-        try (final Connection connection = DataSourceUtils.getConnection(dataSource);
-             final PreparedStatement pstmt = connection.prepareStatement(sql)) {
-            log.debug("query : {}", sql);
-
+    public <T> List<T> query(final String sql, final RowMapper<T> rowMapper, final Object... args) {
+        return preparedStatementExecutor.execute(sql, pstmt -> {
             ResultSet resultSet = pstmt.executeQuery();
             final List<T> results = new ArrayList<>();
             while (resultSet.next()) {
                 results.add(rowMapper.mapRow(resultSet, resultSet.getRow()));
             }
-
             return results;
-        } catch (SQLException e) {
-            log.error(e.getMessage(), e);
-            throw new DataAccessException(e);
-        }
+        }, args);
     }
 
     public <T> T queryForObject(final String sql, final RowMapper<T> rowMapper, final Object... args) {
-        try (final Connection connection = DataSourceUtils.getConnection(dataSource);
-             final PreparedStatement pstmt = connection.prepareStatement(sql)) {
-            log.debug("query : {}", sql);
-
-            setPreparedStatementArguments(pstmt, args);
+        return preparedStatementExecutor.execute(sql, pstmt -> {
             ResultSet resultSet = pstmt.executeQuery();
             final List<T> results = new ArrayList<>();
             while (resultSet.next()) {
@@ -68,16 +40,7 @@ public class JdbcTemplate {
             }
             validateResultSize(results);
             return results.get(0);
-        } catch (SQLException e) {
-            log.error(e.getMessage(), e);
-            throw new DataAccessException(e);
-        }
-    }
-
-    private void setPreparedStatementArguments(final PreparedStatement pstmt, final Object[] args) throws SQLException {
-        for (int parameterIndex = 1; parameterIndex <= args.length; parameterIndex++) {
-            pstmt.setObject(parameterIndex, args[parameterIndex - 1]);
-        }
+        }, args);
     }
 
     private <T> void validateResultSize(final List<T> results) {
@@ -89,5 +52,4 @@ public class JdbcTemplate {
                     String.format("Incorrect result size : expected - %d, actual - %d", 1, results.size()));
         }
     }
-
 }
