+++ jdbc/src/main/java/org/springframework/jdbc/core/PreparedStatementCallback.java
@@ -1,0 +1,10 @@
+package org.springframework.jdbc.core;
+
+import java.sql.PreparedStatement;
+import java.sql.SQLException;
+
+@FunctionalInterface
+public interface PreparedStatementCallback<T> {
+
+    T doInPreparedStatement(PreparedStatement pstmt) throws SQLException;
+}
